requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=2.0.0
python-dotenv>=1.0.0
rich>=13.7.0
ta>=0.11.0
yfinance>=0.2.40
beautifulsoup4>=4.12.3
lxml>=5.1.0
schedule>=1.2.1
colorama>=0.4.6
streamlit>=1.32.0
plotly>=5.18.0
groq>=0.9.0
